+++ packages/button/Button.tsx
@@ -1,9 +1,15 @@
 import { styled } from '@pismo/bolt-core'
+import * as React from 'react'
 
 import cleanButton from './polished/cleanButton'
 
 export interface TypeButtonProps {
+  /**
+   * Set the button color theme to secondary
+   * @default false
+   */
   secondary?: boolean
+  children?: React.ReactNode
 }
 
 const getColorByType = (props: TypeButtonProps) => {
@@ -23,7 +29,7 @@ const getColorByType = (props: TypeButtonProps) => {
   return colors.default
 }
 
-export const Button = styled.button<TypeButtonProps>`
+const StyledButton = styled.button<TypeButtonProps>`
   ${cleanButton()};
   border-radius: 4px;
   background-color: ${props => getColorByType(props).backgroundColor};
@@ -40,3 +46,7 @@ export const Button = styled.button<TypeButtonProps>`
     background-color: ${props => getColorByType(props).backgroundColorHover};
   }
 `
+
+export const Button = (props: TypeButtonProps & React.HTMLAttributes<HTMLButtonElement>) => (
+  <StyledButton {...props}>{props.children}</StyledButton>
+)