+++ packages/button/polished/cleanButton.d.ts
@@ -1,0 +1,8 @@
+declare const _default: () => {
+  background: string
+  border: string
+  padding: number
+  cursor: string
+  outline: string
+}
+export default _default