+++ packages/form-section/FormSection.stories.tsx
@@ -1,0 +1,6 @@
+import { storiesOf } from '@storybook/react'
+import * as React from 'react'
+
+import { FormSection } from './FormSection'
+
+storiesOf('Components/FormSection', module).add('Usage', () => <FormSection> My FormSection </FormSection>)