+++ packages/form-section/FormSection.tsx
@@ -1,6 +1,7 @@
 import { Flex, media, styled } from '@pismo/bolt-core'
+import * as React from 'react'
 
-export const FormSection = styled(Flex)`
+export const StyledFormSection = styled(Flex)`
   justify-content: space-between;
 
   ${media.mobile`
@@ -8,3 +9,9 @@ export const FormSection = styled(Flex)`
     flex-direction: column;
   `}
 `
+
+export interface FormSectionProps {
+  children?: React.ReactNode
+}
+
+export const FormSection = (props: FormSectionProps) => <StyledFormSection> {props.children} </StyledFormSection>