+++ packages/form-section/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/pismo/bolt/compare/v0.6.0...v0.7.0) (2019-04-17)
+
+
+### Features
+
+* **Storybook:** add more stories ([b54a57a](https://github.com/pismo/bolt/commit/b54a57a))
+
+
+
+
+
 # [0.6.0](https://github.com/pismo/bolt/compare/v0.5.0...v0.6.0) (2019-04-14)
 
 **Note:** Version bump only for package @pismo/bolt-form-section