+++ packages/select-field/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/pismo/bolt/compare/v0.6.0...v0.7.0) (2019-04-17)
+
+
+### Bug Fixes
+
+* **Storybook:** Button, Loader and TableItem are now able to receive HTML props again; fixes to SelectField and TextField ([980a2b6](https://github.com/pismo/bolt/commit/980a2b6))
+* commited wrong files in last commit ([95c7b98](https://github.com/pismo/bolt/commit/95c7b98))
+
+
+### Features
+
+* **Storybook:** add more stories ([b54a57a](https://github.com/pismo/bolt/commit/b54a57a))
+
+
+
+
+
 # [0.6.0](https://github.com/pismo/bolt/compare/v0.5.0...v0.6.0) (2019-04-14)